Stabilize process-tree timeout coverage
