+++ src/adapters/process/mod.rs
@@ -330,138 +330,4 @@ fn finish_worker<T>(worker: &mut Worker<T>, deadline: Deadline) -> Result<(), Pr
 }
 
 #[cfg(test)]
-mod tests {
-    use std::ffi::OsString;
-
-    #[cfg(unix)]
-    use std::time::Duration;
-
-    use crate::{adapters::memory::FakeIdGenerator, ports::environment::IdGenerator as _};
-
-    #[cfg(unix)]
-    use crate::ports::environment::{ProcessError, ProcessRequest, ProcessRunner};
-
-    use super::resume_args;
-
-    #[cfg(unix)]
-    use super::SystemProcessRunner;
-
-    #[test]
-    fn replacement_resume_arguments_preserve_an_explicit_state_root() {
-        let mut ids = FakeIdGenerator::new(1_800_000_000_000);
-        let session = ids.session_id();
-        assert_eq!(
-            resume_args(session, Some(std::path::Path::new("/private/state"))),
-            ["--state-dir", "/private/state", "-r", &session.to_string(),].map(OsString::from)
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn arguments_and_standard_input_are_not_shell_interpolated() {
-        let mut runner = SystemProcessRunner::default();
-        let output = runner
-            .run(ProcessRequest {
-                program: OsString::from("/bin/cat"),
-                args: Vec::new(),
-                stdin: Some(b"$(touch never) ; exact\n".to_vec()),
-                timeout: Duration::from_secs(1),
-            })
-            .expect("direct process");
-        assert_eq!(output.exit_code, Some(0));
-        assert_eq!(output.stdout, b"$(touch never) ; exact\n");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn deadline_terminates_a_slow_process() {
-        let mut runner = SystemProcessRunner::default();
-        let result = runner.run(ProcessRequest {
-            program: OsString::from("/bin/sleep"),
-            args: vec![OsString::from("2")],
-            stdin: None,
-            timeout: Duration::from_millis(10),
-        });
-        assert_eq!(result, Err(ProcessError::TimedOut));
-        let recovery = runner
-            .run(ProcessRequest {
-                program: OsString::from("/bin/cat"),
-                args: Vec::new(),
-                stdin: Some(b"recovered".to_vec()),
-                timeout: Duration::from_secs(1),
-            })
-            .expect("later process remains usable");
-        assert_eq!(recovery.stdout, b"recovered");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn shared_cancellation_terminates_running_process_work() {
-        let cancellation = super::CancellationFlag::default();
-        let worker_cancellation = cancellation.clone();
-        let handle = std::thread::spawn(move || {
-            let mut runner = SystemProcessRunner::cancellable(worker_cancellation);
-            runner.run(ProcessRequest {
-                program: OsString::from("/bin/sleep"),
-                args: vec![OsString::from("30")],
-                stdin: None,
-                timeout: Duration::from_secs(30),
-            })
-        });
-        std::thread::sleep(Duration::from_millis(30));
-        let started = std::time::Instant::now();
-        cancellation.cancel();
-        let result = handle.join().expect("process runner thread");
-        assert_eq!(result, Err(ProcessError::Cancelled));
-        assert!(started.elapsed() < Duration::from_secs(1));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn inherited_output_pipe_cannot_extend_the_deadline() {
-        let started = std::time::Instant::now();
-        let mut runner = SystemProcessRunner::default();
-        let result = runner.run(ProcessRequest {
-            program: OsString::from("/bin/sh"),
-            args: vec![OsString::from("-c"), OsString::from("sleep 2 &")],
-            stdin: None,
-            timeout: Duration::from_millis(30),
-        });
-        assert_eq!(result, Err(ProcessError::TimedOut));
-        assert!(started.elapsed() < Duration::from_millis(500));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn timeout_terminates_a_grandchild_that_inherits_output_pipes() {
-        let temporary = tempfile::tempdir().expect("temporary fixture");
-        let pid_file = temporary.path().join("grandchild.pid");
-        let script = "sleep 30 & child=$!; echo $child > \"$1\"; wait";
-        let mut runner = SystemProcessRunner::default();
-        let result = runner.run(ProcessRequest {
-            program: OsString::from("/bin/sh"),
-            args: vec![
-                OsString::from("-c"),
-                OsString::from(script),
-                OsString::from("proqi-process-fixture"),
-                pid_file.as_os_str().to_owned(),
-            ],
-            stdin: None,
-            timeout: Duration::from_millis(50),
-        });
-        assert_eq!(result, Err(ProcessError::TimedOut));
-        let raw = std::fs::read_to_string(pid_file)
-            .expect("grandchild pid")
-            .trim()
-            .parse::<i32>()
-            .expect("numeric pid");
-        let pid = rustix::process::Pid::from_raw(raw).expect("positive pid");
-        for _ in 0..50 {
-            if rustix::process::test_kill_process(pid).is_err() {
-                return;
-            }
-            std::thread::sleep(Duration::from_millis(10));
-        }
-        panic!("grandchild survived process-tree cancellation");
-    }
-}
+mod tests;
